refactor: merge pandas
