--- src/openaivec/aio/util.py
@@ -1,6 +1,0 @@
-from typing import TypeVar
-
-__all__ = []
-
-T = TypeVar("T")
-U = TypeVar("U")
